Supervisord: redirect proc logs to `/dev/stdout`
Supervisord manages the underlying clickhouse and plausible server
processes. In this commit, we change the logging directives to have
supervisord redirect the log output to `/dev/stdout` and `/dev/stderr`,
which are captured by Docker. This way we can see the process logs
directly within the native Cloudron log console. See:
https://veithen.io/2015/01/08/supervisord-redirecting-stdout.html