Update `README.md` with further to-do tasks
Detailed inspection of logging reveals further read/write access issues
with some of Plausible's libraries. Tzdata is used for time-zone data,
and the previous set of patches and symlinks did not completely resolve
the EACCESS problem. Locus is used for geolocation with the MaxMind
GeoIP databases, and likewise requires write access for caching.