Add `PERSISTENT_CACHE_DIR` Plausible Env
`PERSISTENT_CACHE_DIR` is an undocumented Plausible environment variable
that is currently only used to define a location for the Locus
geolocation library to store it's geolocation database cache. I
encountered this option when debugging the locus geolocation database
cache access issue.

This commit adds a new directory at `/run/plausible/cache_dir`, and then
sets the `PERSISTENT_CACHE_DIR` environment variable to point to it.

For further information see:
https://github.com/plausible/analytics/blob/af6b578dc5dce94ec0bac6ab31f4be5bd8007ac3/config/runtime.exs#L232