Define clickhouse paths, add client configuration
This commit explicitly defines as many of Clickhouse server's data paths
as possible in `./configs/clickhouse-config.xml.template`, in
preparation for a future change to how clickhouse is run. Right now the
primary clickhouse data directory is set to `/app/data/clickhouse/`, but
we will hope to change it to `/run/clickhouse/` in the future once we
implement proper database backup and restore functionality. The
long-term goal is to have clickhouse back itself up to
/app/data/clickhouse/`, and load tables from backup on app start.

The second major change is the inclusion of a `clickhouse-client.xml`
file that is used to define the connection settings for the command line
clickhouse client. We will be using `clickhouse-client` to perform the
database backups in the future.