Impl. backup and restore scripts for clickhouse
This is a substantive commit which implements backup and restore scripts
for the clickhouse database. This implementation is a part of a broader
effort towards setting the clickhouse data directory (the `<path>` tag
in `clickhouse-config.xml` to the `/run/clickhouse` directory, in order
to keep persistent data in the form of database backups in `/app/data`.

In order to implement backups, we first define a backup location at
`/app/data/clickhouse/backups` in `clickhouse-config.xml`. We then
create a pair of bash scripts that backup and restore to said location
using `clickhouse-client`. The `clickhouse-backup.sh` script allows for
incremental backups, although currently this feature is disabled as it
does not seem to save much file size.

We also modify our `start.sh` file to take a backup of our clickhouse
database immediately after first installation. This commit also includes
a few final misc. changes to `start.sh`, mostly to make it more robust
and organized.