Move code from `start.sh` to `initial-setup.sh`
This commit refactors the code in `start.sh` for greater readability and
ease-of-maintenance. The original `start.sh` startup code was very
bloated and convoluted, because it also contains the code for initial
setup on first installation. This code, which is for generating secret
keys, and initializing the plausible database, is not needed for initial
startup and confuses readers. This commit moves all of it out into its
own file at `initial-setup.sh` and also adds minor readability
improvements to `start.sh`.

A major addition of code can also be found in the
`clickhouse-dataloader.sh` file, which defines a dataloader service that
is supposed to load the plausible_events_db data from backup upon
application startup, and back up the live data into
`/app/data/plausible` upon app shutdown. This code was added in order to
move the entire clickhouse data dir to `/run/clickhouse`, in hopes of a
future implementation where the database is only stored as backups in
the persistent directory. This is because it is not safe to backup a
database directly by it's runtime directory.

Unfortunately, I've encountered issues with the script being unable to
receive (?) unix signals from supervisord. So the code does not work,
hence its relevent supervisord entry is commented out. I hope to find a
solution for this in the future.

Finally, this commit also cleans up the `Dockerfile`, and moves some
hard-coded paths into `ARG` variables for easier maintenance in the
future.